requirements-mkl.txt
intel-numpy
intel-scipy
pytest
tensornetwork==0.4.0
